toolkit/components/passwordmgr/test/mochitest/test_insecure_form_field_no_saved_login.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test basic login, contextual inscure password warning without saved logins</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/AddTask.js"></script>
  <script type="text/javascript" src="satchel_common.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: contextual inscure password warning without saved logins

<script>
let chromeScript = runChecksAfterCommonInit();
</script>
<p id="display"></p>

<!-- we presumably can't hide the content for this test. -->
<div id="content">

  <form id="form1" action="http://autocomplete:8888/formtest.js" onsubmit="return false;">
    <input  type="text"       name="uname">
    <input  type="password"   name="pword">
    <button type="submit">Submit</button>
  </form>

</div>

<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Login Manager: contextual inscure password warning without saved logins. **/

// Set to pref before the document loads.
SpecialPowers.setBoolPref(
  "security.insecure_field_warning.contextual.enabled", true);

SimpleTest.registerCleanupFunction(() => {
  SpecialPowers.clearUserPref(
    "security.insecure_field_warning.contextual.enabled");
});

let uname = $_(1, "uname");
let pword = $_(1, "pword");

// Restore the form to the default state.
function restoreForm() {
  uname.value = "";
  pword.value = "";
  uname.focus();
}

// Check for expected username/password in form.
function checkACForm(expectedUsername, expectedPassword) {
  let formID = uname.parentNode.id;
  is(uname.value, expectedUsername, "Checking " + formID + " username is: " + expectedUsername);
  is(pword.value, expectedPassword, "Checking " + formID + " password is: " + expectedPassword);
}

function spinEventLoop() {
  return Promise.resolve();
}

add_task(async function setup() {
  listenForUnexpectedPopupShown();
});

add_task(async function test_form1_initial_empty() {
  await SimpleTest.promiseFocus(window);

  // Make sure initial form is empty.
  checkACForm("", "");
  let popupState = await getPopupState();
  is(popupState.open, false, "Check popup is initially closed");
});

add_task(async function test_form1_warning_entry() {
  await SimpleTest.promiseFocus(window);
  // Trigger autocomplete popup
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  let popupState = await getPopupState();
  is(popupState.open, true, "Check popup is opened");
  is(popupState.selectedIndex, -1, "Check no entries are selected upon opening");

  synthesizeKey("KEY_ArrowDown"); // select insecure warning
  checkACForm("", ""); // value shouldn't update just by selecting
  synthesizeKey("KEY_Enter");
  await spinEventLoop(); // let focus happen
  checkACForm("", "");
});

</script>
</pre>
</body>
</html>
